backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-google-genai
qdrant-client
sentence-transformers
pydantic

langchain-community
